Refactoring OpenShiftRecorderHandler in order to split recording logic into a separate service
